First attempt: count "questions out of order"
Sort questions from "smallest number of people that got it right" to "largest
number of people that got it right" (hardest to easiest)

Then, for each player, count the number of swaps you'd have to do to get all
that player's wrong answers to come before all of that player's right answers
(when questions are ordered hardest to easiest). Guess that the player with the
largest number of swaps is the cheater.

This doesn't work, as it doesn't account for swapping a lot of
similar-difficulty questions vs correctly getting a few very unlikely correct
answers